+++ backend/src/apiserver/list/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//backend/api:go_default_library",
+        "//backend/src/apiserver/common:go_default_library",
         "//backend/src/apiserver/filter:go_default_library",
         "//backend/src/common/util:go_default_library",
         "@com_github_masterminds_squirrel//:go_default_library",
@@ -19,10 +20,12 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//backend/api:go_default_library",
+        "//backend/src/apiserver/common:go_default_library",
         "//backend/src/apiserver/filter:go_default_library",
         "//backend/src/common/util:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
         "@com_github_google_go_cmp//cmp/cmpopts:go_default_library",
         "@com_github_masterminds_squirrel//:go_default_library",
+        "@com_github_stretchr_testify//assert:go_default_library",
     ],
 )
