+++ backend/src/apiserver/list/list.go
@@ -18,6 +18,7 @@
 package list
 
 import (
+	"database/sql"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
@@ -26,6 +27,7 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	api "github.com/kubeflow/pipelines/backend/api/go_client"
+	"github.com/kubeflow/pipelines/backend/src/apiserver/common"
 	"github.com/kubeflow/pipelines/backend/src/apiserver/filter"
 	"github.com/kubeflow/pipelines/backend/src/common/util"
 )
@@ -198,6 +200,42 @@ func (o *Options) AddFilterToSelect(sqlBuilder sq.SelectBuilder) sq.SelectBuilde
 	}
 
 	return sqlBuilder
+}
+
+// FilterOnResourceReference filters the given resource's table by rows from the ResourceReferences
+// table that match an optional given filter, and returns the rebuilt SelectBuilder
+func FilterOnResourceReference(tableName string, resourceType common.ResourceType, selectCount bool,
+	filterContext *common.FilterContext) (sq.SelectBuilder, error) {
+	selectBuilder := sq.Select("*")
+	if selectCount {
+		selectBuilder = sq.Select("count(*)")
+	}
+	selectBuilder = selectBuilder.From(tableName)
+	if filterContext.ReferenceKey != nil {
+		resourceReferenceFilter, args, err := sq.Select("ResourceUUID").
+			From("resource_references as rf").
+			Where(sq.And{
+				sq.Eq{"rf.ResourceType": resourceType},
+				sq.Eq{"rf.ReferenceUUID": filterContext.ID},
+				sq.Eq{"rf.ReferenceType": filterContext.Type}}).ToSql()
+		if err != nil {
+			return selectBuilder, util.NewInternalServerError(
+				err, "Failed to create subquery to filter by resource reference: %v", err.Error())
+		}
+		return selectBuilder.Where(fmt.Sprintf("UUID in (%s)", resourceReferenceFilter), args...), nil
+	}
+	return selectBuilder, nil
+}
+
+// Scans the one given row into a number, and returns the number
+func ScanRowToTotalSize(rows *sql.Rows) (int, error) {
+	var total_size int
+	rows.Next()
+	err := rows.Scan(&total_size)
+	if err != nil {
+		return 0, util.NewInternalServerError(err, "Failed to scan row total_size")
+	}
+	return total_size, nil
 }
 
 // Listable is an interface that should be implemented by any resource/model
