+++ backend/src/apiserver/storage/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "db_status_store.go",
         "experiment_store.go",
         "job_store.go",
-        "list_util.go",
         "minio_client.go",
         "minio_client_fake.go",
         "object_store.go",
@@ -47,7 +46,6 @@ go_test(
         "db_test.go",
         "experiment_store_test.go",
         "job_store_test.go",
-        "list_util_test.go",
         "object_store_test.go",
         "pipeline_store_test.go",
         "resource_reference_store_test.go",
